+++ sharding-core/sharding-core-rewrite/src/test/java/org/apache/shardingsphere/core/rewrite/feature/encrypt/token/pojo/EncryptLiteralAssignmentTokenTest.java
@@ -15,9 +15,8 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.core.rewrite.sql.token;
+package org.apache.shardingsphere.core.rewrite.feature.encrypt.token.pojo;
 
-import org.apache.shardingsphere.core.rewrite.feature.encrypt.token.pojo.EncryptLiteralAssignmentToken;
 import org.junit.Test;
 
 import static org.hamcrest.CoreMatchers.is;
