+++ shardingsphere-sql-rewriter/src/main/java/org/apache/shardingsphere/core/rewrite/sql/token/generator/CollectionSQLTokenGenerator.java
